# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RestaurantImportPlanInput {
  /**
     * @minItems 1
     * @maxItems 7
     * @items.minLength 2
     */
  cities: string[];
  /**
     * @minimum 1
     * @maximum 20
     */
  perCityLimit?: number;
  /** @minimum 1 */
  monthlyBudgetCents: number;
}

export type RestaurantImportRunInput = RestaurantImportPlanInput & {
  confirm: boolean;
};

export type RestaurantImportCityPlanStatus = typeof RestaurantImportCityPlanStatus[keyof typeof RestaurantImportCityPlanStatus];


export const RestaurantImportCityPlanStatus = {
  ready: 'ready',
  capped: 'capped',
  skipped: 'skipped',
} as const;

export interface RestaurantImportCityPlan {
  city: string;
  requested: number;
  estimatedApiCalls: number;
  estimatedCostCents: number;
  status: RestaurantImportCityPlanStatus;
}

export interface RestaurantImportPlan {
  cities: RestaurantImportCityPlan[];
  totalRestaurants: number;
  totalApiCalls: number;
  estimatedCostCents: number;
  monthlyBudgetCents: number;
  withinBudget: boolean;
  note: string;
}

export interface RestaurantImportStatus {
  monthlyBudgetCents: number;
  spentCents: number;
  remainingCents: number;
  callsUsed: number;
  restaurantsImported: number;
  /** @nullable */
  lastRunAt: string | null;
  cities: string[];
}

export interface Restaurant {
  id: string;
  name: string;
  address: string;
  city: string;
  /** @nullable */
  rating: number | null;
  /** @nullable */
  website: string | null;
  googleMapsUrl: string;
  types: string[];
  outreachStatus: string;
  claimed: boolean;
}

export interface RestaurantClaimInput {
  /**
     * @maxLength 254
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email: string;
}

export type RestaurantClaimResultStatus = typeof RestaurantClaimResultStatus[keyof typeof RestaurantClaimResultStatus];


export const RestaurantClaimResultStatus = {
  pending_checkout: 'pending_checkout',
} as const;

export interface RestaurantClaimResult {
  placeId: string;
  attemptId: string;
  status: RestaurantClaimResultStatus;
  monthlyPricePence: 9900;
  currency: 'gbp';
}

export interface RestaurantCheckoutInput {
  /**
     * @maxLength 254
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email: string;
  attemptId: string;
}

export interface RestaurantCheckoutResult {
  /** @pattern ^https:// */
  checkoutUrl: string;
}

export type CheckoutCompletionClaimStatus = typeof CheckoutCompletionClaimStatus[keyof typeof CheckoutCompletionClaimStatus];


export const CheckoutCompletionClaimStatus = {
  processing: 'processing',
  pending_checkout: 'pending_checkout',
  checkout_created: 'checkout_created',
  active: 'active',
  past_due: 'past_due',
  revoked: 'revoked',
} as const;

export type CheckoutCompletionPaymentStatus = typeof CheckoutCompletionPaymentStatus[keyof typeof CheckoutCompletionPaymentStatus];


export const CheckoutCompletionPaymentStatus = {
  processing: 'processing',
  paid: 'paid',
  unpaid: 'unpaid',
  no_payment_required: 'no_payment_required',
} as const;

export interface CheckoutCompletion {
  /** @nullable */
  placeId: string | null;
  /** @nullable */
  name: string | null;
  claimStatus: CheckoutCompletionClaimStatus;
  paymentStatus: CheckoutCompletionPaymentStatus;
}

export interface UnsubscribeResult {
  suppressed: boolean;
}

export interface IncomingReplyInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  placeId: string;
  /**
     * @minLength 1
     * @maxLength 10000
     */
  body: string;
  /**
     * @maxLength 254
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  from?: string;
}

export type ReplyClassificationCategory = typeof ReplyClassificationCategory[keyof typeof ReplyClassificationCategory];


export const ReplyClassificationCategory = {
  unsubscribe: 'unsubscribe',
  wrong_contact: 'wrong_contact',
  out_of_office: 'out_of_office',
  not_interested: 'not_interested',
  upgrade: 'upgrade',
  interested: 'interested',
  question: 'question',
  unknown: 'unknown',
} as const;

export type ReplyClassificationConfidence = typeof ReplyClassificationConfidence[keyof typeof ReplyClassificationConfidence];


export const ReplyClassificationConfidence = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type ReplyClassificationRecommendedAction = typeof ReplyClassificationRecommendedAction[keyof typeof ReplyClassificationRecommendedAction];


export const ReplyClassificationRecommendedAction = {
  suppress: 'suppress',
  pause: 'pause',
  sales_follow_up: 'sales_follow_up',
  answer_question: 'answer_question',
  manual_review: 'manual_review',
} as const;

export interface ReplyClassification {
  category: ReplyClassificationCategory;
  confidence: ReplyClassificationConfidence;
  /** @nullable */
  matchedPattern: string | null;
  recommendedAction: ReplyClassificationRecommendedAction;
}

export interface GmailSyncResult {
  /**
     * @minimum 0
     * @maximum 50
     */
  searchedThreads: number;
  /**
     * @minimum 0
     * @maximum 50
     */
  matchedThreads: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  processed: number;
  /** @minimum 0 */
  skipped: number;
  /** @minimum 0 */
  failed: number;
  capped: boolean;
  /**
     * @maxItems 20
     * @items.maxLength 200
     */
  errors: string[];
}

export type RestaurantEnrichmentResultConfidence = typeof RestaurantEnrichmentResultConfidence[keyof typeof RestaurantEnrichmentResultConfidence];


export const RestaurantEnrichmentResultConfidence = {
  none: 'none',
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface RestaurantEnrichmentResult {
  message: string;
  placeId: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  finalUrl: string | null;
  cuisines: string[];
  dietaryTags: string[];
  confidence: RestaurantEnrichmentResultConfidence;
}

export interface OutreachRunResult {
  /** @minimum 0 */
  discovered: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  sent: number;
  /** @minimum 0 */
  skipped: number;
  /** @minimum 0 */
  failed: number;
  dailyMaximum: 20;
}

export interface RestaurantImportRunResult {
  imported: number;
  skippedDuplicates: number;
  apiCalls: number;
  chargedCents: number;
  stoppedBecause: string;
  restaurants: Restaurant[];
}

export type ListRestaurantsParams = {
city?: string;
};