# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: restaurants
    description: Budget-safe restaurant discovery and import operations
  - name: subscriptions
    description: Restaurant claim and subscription operations
  - name: outreach
    description: Outreach preferences and automation
  - name: replies
    description: Protected inbound outreach reply classification
  - name: gmail
    description: Protected Gmail reply polling
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /restaurant-import/status:
    get:
      operationId: getRestaurantImportStatus
      tags: [restaurants]
      summary: Get restaurant import budget status
      responses:
        "200":
          description: Current budget and import status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantImportStatus"
  /restaurant-import/plan:
    post:
      operationId: createRestaurantImportPlan
      tags: [restaurants]
      summary: Preview a budget-safe import plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantImportPlanInput"
      responses:
        "200":
          description: Import plan preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantImportPlan"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /restaurant-import/run:
    post:
      operationId: runRestaurantImport
      tags: [restaurants]
      summary: Import restaurants within a monthly budget cap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantImportRunInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantImportRunResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Google Places is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List imported restaurants
      parameters:
        - in: query
          name: city
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Imported restaurants, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"
  /restaurants/{placeId}/claim:
    post:
      operationId: claimRestaurant
      tags: [subscriptions]
      summary: Start a restaurant ownership claim
      parameters:
        - $ref: "#/components/parameters/PlaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantClaimInput"
      responses:
        "200":
          description: Claim recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantClaimResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /restaurants/{placeId}/checkout:
    post:
      operationId: createRestaurantCheckout
      tags: [subscriptions]
      summary: Create a £99 GBP monthly Stripe Checkout session
      parameters:
        - $ref: "#/components/parameters/PlaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantCheckoutInput"
      responses:
        "200":
          description: Stripe-hosted checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantCheckoutResult"
        "400":
          description: Invalid request or price configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkout-completion/{sessionId}:
    get:
      operationId: getCheckoutCompletion
      tags: [subscriptions]
      summary: Get the minimal status of a completed Stripe Checkout session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 255
      responses:
        "200":
          description: Checkout completion status (may briefly be processing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutCompletion"
        "400":
          description: Invalid session ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Stripe or public URL is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /outreach/unsubscribe/{token}:
    post:
      operationId: unsubscribeRestaurantOutreach
      tags: [outreach]
      summary: Suppress all future outreach for a restaurant
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            minLength: 32
      responses:
        "200":
          description: Suppression recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsubscribeResult"
        "404":
          description: Invalid unsubscribe token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /automation/outreach:
    post:
      operationId: runOutreachAutomation
      tags: [outreach]
      summary: Run one protected, daily outreach batch (maximum 20 sends)
      security:
        - automationBearer: []
      responses:
        "200":
          description: Automation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachRunResult"
        "401":
          description: Invalid automation credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Sending is disabled or configuration is unsafe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /replies/incoming:
    post:
      operationId: classifyIncomingReply
      tags: [replies]
      summary: Classify a supplied outreach reply and update its restaurant
      description: Classifies only the supplied body. It does not read or modify an email mailbox.
      security:
        - automationBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomingReplyInput"
      responses:
        "200":
          description: Reply classified and recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyClassification"
        "400":
          description: Invalid or oversized request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid automation credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gmail/sync:
    post:
      operationId: syncGmailReplies
      tags: [gmail]
      summary: Poll Gmail for replies to previously recorded outreach threads
      description: Read-only bounded polling. Gmail messages are not modified.
      security:
        - automationBearer: []
      responses:
        "200":
          description: Gmail polling result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GmailSyncResult"
        "401":
          description: Invalid automation credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Gmail polling could not start
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /restaurants/{placeId}/enrich:
    post:
      operationId: enrichRestaurant
      tags: [restaurants]
      summary: Enrich one restaurant from its public website
      security:
        - automationBearer: []
      parameters:
        - $ref: "#/components/parameters/PlaceId"
      responses:
        "200":
          description: Enrichment completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantEnrichmentResult"
        "401":
          description: Invalid automation credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Restaurant cannot be enriched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    automationBearer:
      type: http
      scheme: bearer
  parameters:
    PlaceId:
      in: path
      name: placeId
      required: true
      schema:
        type: string
        minLength: 1
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RestaurantImportPlanInput:
      type: object
      properties:
        cities:
          type: array
          items:
            type: string
            minLength: 2
          minItems: 1
          maxItems: 7
        perCityLimit:
          type: number
          multipleOf: 1
          minimum: 1
          maximum: 20
          default: 10
        monthlyBudgetCents:
          type: number
          multipleOf: 1
          minimum: 1
      required:
        - cities
        - monthlyBudgetCents
    RestaurantImportRunInput:
      allOf:
        - $ref: "#/components/schemas/RestaurantImportPlanInput"
        - type: object
          properties:
            confirm:
              type: boolean
          required:
            - confirm
    RestaurantImportCityPlan:
      type: object
      properties:
        city:
          type: string
        requested:
          type: number
          multipleOf: 1
        estimatedApiCalls:
          type: number
          multipleOf: 1
        estimatedCostCents:
          type: number
          multipleOf: 1
        status:
          type: string
          enum: [ready, capped, skipped]
      required:
        - city
        - requested
        - estimatedApiCalls
        - estimatedCostCents
        - status
    RestaurantImportPlan:
      type: object
      properties:
        cities:
          type: array
          items:
            $ref: "#/components/schemas/RestaurantImportCityPlan"
        totalRestaurants:
          type: number
          multipleOf: 1
        totalApiCalls:
          type: number
          multipleOf: 1
        estimatedCostCents:
          type: number
          multipleOf: 1
        monthlyBudgetCents:
          type: number
          multipleOf: 1
        withinBudget:
          type: boolean
        note:
          type: string
      required:
        - cities
        - totalRestaurants
        - totalApiCalls
        - estimatedCostCents
        - monthlyBudgetCents
        - withinBudget
        - note
    RestaurantImportStatus:
      type: object
      properties:
        monthlyBudgetCents:
          type: number
          multipleOf: 1
        spentCents:
          type: number
          multipleOf: 1
        remainingCents:
          type: number
          multipleOf: 1
        callsUsed:
          type: number
          multipleOf: 1
        restaurantsImported:
          type: number
          multipleOf: 1
        lastRunAt:
          type: ["string", "null"]
        cities:
          type: array
          items:
            type: string
      required:
        - monthlyBudgetCents
        - spentCents
        - remainingCents
        - callsUsed
        - restaurantsImported
        - lastRunAt
        - cities
    Restaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        city:
          type: string
        rating:
          type: ["number", "null"]
        website:
          type: ["string", "null"]
        googleMapsUrl:
          type: string
        types:
          type: array
          items:
            type: string
        outreachStatus:
          type: string
        claimed:
          type: boolean
      required:
        - id
        - name
        - address
        - city
        - rating
        - website
        - googleMapsUrl
        - types
        - outreachStatus
        - claimed
    RestaurantClaimInput:
      type: object
      properties:
        email:
          type: string
          pattern: "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$"
          maxLength: 254
      required: [email]
    RestaurantClaimResult:
      type: object
      properties:
        placeId:
          type: string
        attemptId:
          type: string
        status:
          type: string
          enum: [pending_checkout]
        monthlyPricePence:
          type: integer
          const: 9900
        currency:
          type: string
          const: gbp
      required: [placeId, attemptId, status, monthlyPricePence, currency]
    RestaurantCheckoutInput:
      type: object
      properties:
        email:
          type: string
          pattern: "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$"
          maxLength: 254
        attemptId:
          type: string
          format: uuid
      required: [email, attemptId]
    RestaurantCheckoutResult:
      type: object
      properties:
        checkoutUrl:
          type: string
          pattern: "^https://"
      required: [checkoutUrl]
    CheckoutCompletion:
      type: object
      additionalProperties: false
      properties:
        placeId:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        claimStatus:
          type: string
          enum: [processing, pending_checkout, checkout_created, active, past_due, revoked]
        paymentStatus:
          type: string
          enum: [processing, paid, unpaid, no_payment_required]
      required: [placeId, name, claimStatus, paymentStatus]
    UnsubscribeResult:
      type: object
      properties:
        suppressed:
          type: boolean
      required: [suppressed]
    IncomingReplyInput:
      type: object
      additionalProperties: false
      properties:
        placeId:
          type: string
          minLength: 1
          maxLength: 255
        body:
          type: string
          minLength: 1
          maxLength: 10000
        from:
          type: string
          pattern: "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$"
          maxLength: 254
      required: [placeId, body]
    ReplyClassification:
      type: object
      additionalProperties: false
      properties:
        category:
          type: string
          enum: [unsubscribe, wrong_contact, out_of_office, not_interested, upgrade, interested, question, unknown]
        confidence:
          type: string
          enum: [low, medium, high]
        matchedPattern:
          type: ["string", "null"]
        recommendedAction:
          type: string
          enum: [suppress, pause, sales_follow_up, answer_question, manual_review]
      required: [category, confidence, matchedPattern, recommendedAction]
    GmailSyncResult:
      type: object
      additionalProperties: false
      properties:
        searchedThreads:
          type: number
          multipleOf: 1
          minimum: 0
          maximum: 50
        matchedThreads:
          type: number
          multipleOf: 1
          minimum: 0
          maximum: 50
        processed:
          type: number
          multipleOf: 1
          minimum: 0
          maximum: 20
        skipped:
          type: number
          multipleOf: 1
          minimum: 0
        failed:
          type: number
          multipleOf: 1
          minimum: 0
        capped:
          type: boolean
        errors:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 200
      required: [searchedThreads, matchedThreads, processed, skipped, failed, capped, errors]
    RestaurantEnrichmentResult:
      type: object
      properties:
        message:
          type: string
        placeId:
          type: string
        email:
          type: ["string", "null"]
        finalUrl:
          type: ["string", "null"]
        cuisines:
          type: array
          items:
            type: string
        dietaryTags:
          type: array
          items:
            type: string
        confidence:
          type: string
          enum: [none, low, medium, high]
      required:
        - message
        - placeId
        - email
        - finalUrl
        - cuisines
        - dietaryTags
        - confidence
    OutreachRunResult:
      type: object
      properties:
        discovered:
          type: number
          multipleOf: 1
          minimum: 0
        sent:
          type: number
          multipleOf: 1
          minimum: 0
          maximum: 20
        skipped:
          type: number
          multipleOf: 1
          minimum: 0
        failed:
          type: number
          multipleOf: 1
          minimum: 0
        dailyMaximum:
          type: integer
          const: 20
      required: [discovered, sent, skipped, failed, dailyMaximum]
    RestaurantImportRunResult:
      type: object
      properties:
        imported:
          type: number
          multipleOf: 1
        skippedDuplicates:
          type: number
          multipleOf: 1
        apiCalls:
          type: number
          multipleOf: 1
        chargedCents:
          type: number
          multipleOf: 1
        stoppedBecause:
          type: string
        restaurants:
          type: array
          items:
            $ref: "#/components/schemas/Restaurant"
      required:
        - imported
        - skippedDuplicates
        - apiCalls
        - chargedCents
        - stoppedBecause
        - restaurants