# lib/api-zod/src/generated/types/checkoutCompletion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutCompletionClaimStatus } from './checkoutCompletionClaimStatus';
import type { CheckoutCompletionPaymentStatus } from './checkoutCompletionPaymentStatus';

export interface CheckoutCompletion {
  /** @nullable */
  placeId: string | null;
  /** @nullable */
  name: string | null;
  claimStatus: CheckoutCompletionClaimStatus;
  paymentStatus: CheckoutCompletionPaymentStatus;
}