# lib/api-zod/src/generated/types/restaurantCheckoutInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RestaurantCheckoutInput {
  /**
     * @maxLength 254
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email: string;
  attemptId: string;
}